requirements.txt
Django==1.7
django-crispy-forms==1.4.0
django-haystack==2.0.0
Django-Select2==4.2.2
django-tastypie==0.12.1
Pillow==2.7.0
psycopg2==2.6
python-dateutil==2.4.0
python-memcached==1.53
python-mimeparse==0.1.4
requests==2.5.1
simplejson==3.6.5
six==1.9.0
Whoosh==2.4.0
